[inductor] Add restore_value to user-defined Triton kernel support
This PR adds support for the `restore_value` argument of the
`triton.autotune` for the user-defined Triton kernels in PT2.

The `kernel.restore_idx` are extracted in the
`ir.UserDefinedTritonKernel` and the corresponding arg names are
placed into the `triton_meta["restore_value"]`. From there, those
are added to the existing `mutated_arg_names` in the caching autotuner
infra which already exists and leads to the listed argss being cloned.
This achieves the equivalent effect to the native `restore_value`.

ghstack-source-id: a027031e35e3d107c295bdf4f65e901e7e5dc814
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139851